PT: Don't attempt to track whether packet generation is enabled.
PGE and PGD are two kinds of PT packets that mean "Packet Generation
Enable" and "Packet Generation Disable". PT emits these as you enable
and disable tracing.

The decoder tracks whether we are currently in a portion of the trace
where packet generation is enabled. Currently we always use a fresh perf
fd to collect a trace, so we know to start the decoder with
`self.pge=false`.

But going forwards we want to re-use perf fds. In this scenario, when
you collect a trace using a "pre-owned" perf fd, you can't generally
know the state of the packet stream, or whether to start the decoder
with `self.pge=false` or `self.pge=true`.

That sounds problematic, but what I've realised is that whether or not
packet generation is enabled has zero consequence to the decoder, so
there's no need to track it!

This change removes `self.pge` and all of its uses from the decoder. Lo
and behold, all tests still pass.