Resizing the aux buffer causes problems, so don't.
If the collector thread determines that the trace needed more room, it
resizes the buffer and updates the capacity. However, something goes
wrong that causes (at least) the pointer value to become corrupted in
some way. There are several possible candidates: the updates are done
non-atomically in a "collector" thread so may not synchronise with the
being-collected thread; and updates are also not atomically synchronised
with Rust code. My surface-level attempts to fix this haven't succeeded,
so either there's more synchronisation needed than I currently realise,
or there are other problems as well.

This commit thus simply disables trace buffer reallocation. For yk this
isn't likely to be a big problem: we throw away big traces anyway.
[Arguably the current buffer size is much bigger than we need and want,
but that's something we can worry about separately.]

With this commit I can successfully run "yklua db.lua" in a loop without
an error: without this commit, it nearly always dies (normally with a
SEGFAULT in realloc/malloc).