Re-use perf file descriptors for trace collection.
Laurie and Lukas realised that opening Linux perf (`perf_event_open()`)
can be very slow.

To quote Laurie:

> it's bimodal: one peak tightly clustered around 4ms, another peak
> loosely clustered between 20-28ms

This change therefore makes the JIT re-use perf file descriptors
(per-thread).

This required a small change to the state machine in the decoder.
Previously we'd always expected a PSB+ sequence at the start of the
trace, but with a re-used perf fd you won't get one until the next time
`IA32_RTIT_STATUS.PacketByteCnt` exceeds the next emitted-bytes
threshold.

In general Intel recommend that you always scan for a PSB when starting
to decode a trace. We get away without doing this because:

 a) For a new perf fd, the packet stream is always aligned at the start
    of a PSB packet.

 b) Nothing we do causes the packet stream to become desynchronised, so
    for a re-used perf fd, the packet stream is always packet-aligned.

Co-authored-by: Laurence Tratt <laurie@tratt.net>