Handle uncovered control flow events in disassembly-based PT decoding.
There are outstanding 3 cases to handle. Thanks to issue #883, We know
one of them has arisen in the past, but we don't know which.

This change splits out the 3 cases and makes a different crash for each
so that we know which case we are hitting.

I've run this code on try_repeat for 2 hours during which time it
completed 1000 runs of the test suite in release mode, so the crash must
be either no longer present or exceptionally rare.

Let's merge this now, then if it comes up again we can at least know
which case to think about. Depending on which, we may decide to flag a
temporary failure, thus aborting trace compilation and allowing the
interpreter to continue. Cross that bridge if/when we come to it.