Make sure threads don't exit until all LLVM compilation threads complete.
This is a horrible hack: we track in an atomic integer how many worker
threads are doing something (which we use as a proxy for "are doing
something with LLVM) and, when someone calls `yk_mt_drop` in C, we wait
for those worker threads to finish doing whatever they're doing.

The obvious horror with this hack is that we stop the application
exiting even if the interpreter has finished. A less obvious horror is
that `Mt::drop` isn't called while compilation threads are running,
which is why we have to put this hack into `yk_mt_drop` in the C API.
That bakes in an assumption about the C-ness of interpreters, which is
fine for now, but clearly not going forwards.

To emphasise how hacky this is, all the code is under a flag
`yk_llvm_sync_hack`. This will also make it easy to find all the places
to rip out when we no longer need this hack.