Temporarily disable the completion wrapper code.
This has started failing in some situations (the `assert` fails),
including CI, but we don't understand what is really causing that. Since
this is a developer convenience, and not important for the system per
se, this commit temporarily disables the failing code. Someone who
understands this better than I can hopefully reenable it later, but we
don't want to block development in the meantime.