Replace unsound StackAdjust code.
With the recent change to ykllvm, which ensures no calls can appear in the entry
block of a function, we no longer need to rely on the StackAdjust solution in
order to detect callbacks from external functions (i.e. unmappable code).
Instead, we now only need to check if an unmappable block is followed by an
entry block. If it is, then the external function called back into mappable
code; if not then we have returned from the external function.