Start nudging us towards "trace decoding as an iterator".
This commit introduces a `TraceIterator` which in essence replaces
`RawTrace`. Currently the `hwt` backend still does its mapping in a
single step, but it at least exposes an iterator (even if it doesn't
make much use of it).

This commit is clearly a half-way house to what we might eventually
want. In particular it doesn't yet support "trace mapping went wrong
half-way through" because `hwt` can't go wrong in that way -- yet.
But, hopefully, it is a nudge in the right direction.