Rewrite jitmodbuilder inlining/outlining logic.
In order to detect when to inline or outline a function call in a trace,
we previously relied on a complicated call stack that tracked at which
instruction we started outlining (or inlining) and whether we've
processed an unmappable block or not. Returning from a call was also
difficult due to the fact functions calls aren't terminators and we had
to continue tracing within a duplicate block but had to make sure we
continue after the instruction that started outlining. If this sounds
complicated it's because it was. To make matters worse, this logic was
spread through out `jitmodbuilder.cc`.

Luckily, we are now enforcing that 1) no call may appear inside an entry
block and 2) calls are effectively terminators, meaning that there can
only be at most one call per block. This makes it a lot easier to detect
whether we called into or returned from unmappable code. This commit
also moves the entire logic into a single place, making it easier to
reason about and maintain.