Make a start on the JIT IR and the trace builder.
The JIT IR is designed to be small. There are two kinds of instruction:
 - short instructions, with inlined operands.
 - long ones (unimplemented as of yet)

Ideally, function, block and instruction IDs (which are indices read
from the on-disk AOT IR) would be either converted to references as we
decode, but this would require changes to the decoder. We may have to
ditch deku:

https://github.com/sharksforarms/deku/issues/383

Co-authored-by: Lukas Diekmann <lukas.diekmann@gmail.com>