Detect inlined intrinsics by peeking at the trace.
Our previous method of detecting whether an intrinsics was inlined or not
relied on annotating intrinsic calls at compile time. For many intrinsics
this turned out to be very hard if not impossible, depending on how the
intrinsic is lowered.

Since we now only allow one function call per block, this makes detecting
inlined intrinsics at runtime very easy. All we need to do is peek at the
next block in the trace: if it is a mappable block, the intrinsic was inlined;
if the block is unmappable it wasn't inlined (which means we need to copy the
call into the JIT module).