Fix outlining bug.
This change fixes a bug where we forgot to reset the `OutlineBase` in
`jitmodbuilder.cc` after resetting `Outlining` to `false`. This meant we
continued outlining after seeing `yk_mt_control_point`, when we should be
copying into the trace instead. This bug only occured when a call immediately
preceeded the control point and we were tracing at least 2 rounds of the main
interpreter loop. This scenario wasn't previously covered by our test suite.