Use `TiVec` in the AOT IR.
This change replaces auxilliary vectors of the form `Vec<T>` with
`TiVec<T>` from the `typed_index_collections` crate.

`TiVec` is like `IndexVec` from rustc: it uses type-safe indices to
prevent you from mixing incompatible indices by accident.

I looked at a few crates that provide this kind of functionality, and
`typed_index_collections` seemed the best for our needs. Also it appears
to be maintained.

For now we only use `TiVec` for the AOT IR. We may do the same for the
JIT IR later, if we find that it makes sense.

While here, homogonise functions that index in to a vector, making them
panic if the element isn't found (instead of returning `Option`, which I
see no need for at the moment).