Fix a conditional and add some `assert`s to check if we've got UB.
In several cases we ask for elements of a vector (e.g. `index + 1`)
without knowing for sure if those elements exist. It might be that these
extra checks are redundant -- perhaps we know (or once knew) for sure
that they're OK. But since the code doesn't make clear that this is
safe, it's better for us to validate the implicit assumption.