Remove the LLVM JIT compiler ("trace compiler") test suite.
This is a very fragile test suite that percolates throughout the
existing trace compiler in a way that makes it quite hard to make
changes to it (e.g. to support software tracing).

In the short term we want to tweak the LLVM trace compiler to make it easier
to support software tracing. In the medium term we want to remove the
LLVM trace compiler entirely. Fixing the test suite is rather annoying
because of its fragility, and not particularly beneficial to us, so it's
easiest to remove it now and be done with it.

There are probably more bits we can remove and simplify, but this commit
does enough of a job on its own to be a noticeable help.