Encapsulate the trace vector.
Since references to instructions inside a trace are represented by their
index, certain operations, like removal and insertion of instructions
(except appending at the end), are disallowed. In order to make it more
difficult to accidentally do that, this commit encapsulates the vector
storing the instructions of a trace, only exposing operations that are
safe to do.