Propagate errors in the JIT compiler backends.
We create a new error type that expresses that a failure is either
either temporary (i.e. recoverable), or unrecoverable (and we should
crash).

I then replaced all Result return types with a result of this new error
type and audited all unwraps in the new JIT backend, adapting them (where
necessary) to propagate the new error type.

(Since we now know the reason for a failure, we can also print it in
`jit-state` lines)

I've been quite conservative, in that if I'm unsure about if an error
should be temporary, I've classified them as unrecoverable. At least
then we will get a crash and we can think about if we should downgrade
them to temporary or not.

I've not audited the LLVM backend at all. Before this change, any error
in the LLVM backend was considered temporary, and this remains so after
this change. There's no point in auditing the failure modes of the LLVM
backend, because it will be time consuming and we want to delete it
soon anyway.