Pass `TraceIterator` to JIT compilers rather than `Vec<TracedAOTBlock>`.
This allows the JIT compilers, if they so choose, to iterate over the
blocks. If a tracer chooses to process (e.g. decode) its trace in a
thread, this means that (some of) compilation and trace processing can
occur in parallel.

At the moment, this commit doesn't change anything meaningful inside
jitc_ykllvm or jitc_yk: both immediately `collect` the iterator into a
`Vec`. But the API is now flexible enough to allow something better in
the future.