`bb = usize::MAX` isn't a special value.
At some point in the past `bb = usize::MAX` was used to represent
"unmappable" but it's now a separate `enum` item. Nowhere in yk seems to
check/set `bb` for/to `usize::MAX` so this commit removes the
(hopefully) misleading comment. In case we're wrong, I've left behind an
`assert` that's likely to catch any old code. One day far in the future
our successors might feel confident enough to remove the `assert`!