Use `_Generic` to properly select which promote function we want to use.
Although we currently only support `usize`, using `_Generic` now makes
it clearer why we have a macro (I almost removed the macro before I
remembered vaguely why we might have wanted it). `_Generic` is a C11
feature, and there might be interpreters we want to support which don't
use C11, but we can cross that bridge when we come to it.