Process ptradd instructions and global operands.
Add processing of AOT PtrAdd instructions. Since the JIT PtrAdd
instruction is limited to 32 bit offsets, do the conversion here and
return an error in case of a failure so we can abort the trace.

While we are here, handle global operands by emitting a LoadGlobal
instruction and replacing the global operand with an instruction operand
referencing the just emitted instruction.