Make promotion a non-global variable.
Although it's not obvious at first `promote.rs`'s use of a `thread_local!`
is a simple global variable. That is problematic because it means that the
compilation process has to run in the same thread as the tracing thread
which only works if serialised compilation is turned on. If you ran a
program which uses promotion without serialised compilation turned on you
got a "promotion buffer underflowed" panic.

This commit makes the promotion array a non-global variable, so that
serialised compilation is no longer required. This commit makes the change
in the most bone-headed manner possible: as this shows, this isn't a good
fit with our current architecture and makes the structure a lot messier. I
intend addressing this soon but at least this commit unbreaks the system.