Rename trace "collection" to "recording".
[Warning: I will admit that "collection" as a term has never quite
resonated with me, so I'm biased in favour of this commit from the off.]

Trace "collection", to me, implies a slightly passive view of what's
going on, but promotion is a much more active view of what's going on.
In particular, some tracing backends can potentially "decode" (another
term we might want to revisit one day) a trace in such a way that they
successively generate AOT blocks and promotions in one go. That means
that we're going to have to rename the `TracedAOTBlock` enum, probably
to something like `RecordedTraceThingy` where "recorded" might be an AOT
block or a promoted variable (or ... who knows what else in the
future!).

Note that I have chosen not to rename anything in the `hwtracer` crate.
That crate is a bit separate from everything else and I'm not sure I
want to rethink what it's doing at the moment.