Run serialised compilation in another thread.
In retrospect I should have added this to 7eff5624e42bc3f3591b97e424be0748e78bb9bb, but it didn't occur to me at the time.

In essence, to avoid multi-threading bugs, this commit makes testing
serialised compilation more realistic in that compilation is now run in
a different thread. Indeed, we spin up a new thread each time, which is
a slightly tougher test than normal yk (which reuses threads for
compilation). Clearly this wouldn't be a good idea in general, but for
internal testing it should be fine.