Semi-unbreak trace mapping stats.
At some point in refactoring, "trace mapping" costs had been ascribed to
a single `assert` which meant that in `YKD_STATS` we always ended up
with 0 seconds spent in mapping.

In reality, trace mapping is split into two parts. The first (and, at
the moment, the only thing that takes any time) is `stop()` which in hwt
takes a long time to run. So this commit effectively measures the costs
of calling `stop`.

As and when hwt and other backends truly move trace mapping costs to an
iterator, these timing probes will need to be moved again. But at least
I know have a half-decent idea of what the current costs are!