Allow for trace mapper encountering a problem during iteration.
As part of the "truly make mapping an iterator" push, we have to allow
for the possibility that a mapper might only realise there's a problem
with a recorded trace after mapping a portion of it. Thus each `next`
method has the potential to return "there's an error in mapping". This
commit changes `AOTTraceIterator` from:

```rust
pub(crate) trait AOTTraceIterator: Iterator<Item = TraceAction> + Send {}
```

to:

```rust
pub(crate) trait AOTTraceIterator:
    Iterator<Item = Result<TraceAction, AOTTraceIteratorError>> + Send
{}
```

At the moment the `AOTTraceIteratorError` enum is empty, but we now have
the flexibility to add variants easily in the future.