Rethink, and rename, buffer sizes.
I managed to confuse myself and Edd with the original names and sizes of
things. This commit somewhat clarifies things. We've since become (even
more) aware that further work in this part of the system is necessary,
but this commit does at least get us moving, and does so with much
smaller memory allocation than we had previously.