Fix a hidden panics in unit tests.
Upon startup, the system fires off a thread to load the stackmaps in from the
running ELF binary and store them in a LazyLock. This is fine for C programs
that were compiled by ykllvm, but for Rust unit test binaries, there will be
no stackmaps compiled-in, and so the thread panics.

The error isn't detected by the test runner, because we never join() the
thread.

There's no "good place" to join this thread, so this change:
 - a) makes it so the LazyLock initialiser can never panic.
 - b) makes the innards of the LazyLock a `Result`, thus allowing it
      to express failure.

Under this scheme, errors can be detected at the use-sites of the stackmap
data instead of going unnoticed in the un-joined loading thread.

As it happens, no Rust unit tests actually use the stackmaps, so after this
change, loading the stackmaps still fail, just no-one ever looks at the error.