Make hwt trace mapping truly an iterator.
Previously "trace mapping" in hwt:

  1. occurred when `stop()` was called on the tracer.
  2. did everything in one go, creating a `Vec`.

In other words, trace mapping:

  1. stopped the interpreter from running.
  2. didn't allow the compiler to run.

This commit fixes both things by making trace mapping a "true" iterator,
which only does work when `next`. is called [Well, mostly: the first
block is currently mapped during `stop`.] This means that trace mapping
now occurs in the compilation thread and allows the compiler to keep on
running.

We *could* in the future use this to meaningfully put trace mapping (and
even PT decoding) in its own thread(s). I don't want to do that until we
have some measurements that would allow us to tell whether this is worth
the complexity, though.

Overall this commit is fairly cost neutral: it only slightly increases
the code length and has little impact on performance. Note that I have
added two tests (one of which really should have been in the test suite
before; the other of which only really makes sense due to this commit),
which rather increases the size of the commit.