Add `CompiledTrace::as_any` to allow upcasting to `Any`.
Unfortunately, trait upcasting is still
[incomplete/unstable](https://github.com/rust-lang/rust/issues/65991).
Forcing every implementation of `CompiledTrace` to implement `as_any` is
a manual way of allow upcasting.

Counter-intuitively, we use the upcast to `Any` to allow us to downcast
to a concrete implementation of `CompiledTrace`. `depot.rs` can now call
`downcast::<LLVMCompiledTrace>` to get at the `LLVMCompiledTrace`.

This will then allow us to push methods out of `CompiledTrace` and into
`LLVMCompiledTrace`.