Handle globals using normal load/store instructions.
In order to load a global we need to emit two instructions: first we
load the pointer for the global, and then use that pointer to store into
or load from the global.

The AOT instruction `Load(Global("someglobal"))` would result in the
following two JIT instructions:

```
%1 = LoadGlobal("someglobal")
%2 = Load(%1)
```

This means we no longer need separate Load/Store instructions for
globals and can instead share the instructions used for other operands,
e.g. locals.