Remove pointless `unsafe` blocks in an `unsafe` function.
In an `unsafe` function, everything is in a sense unsafe, so these
blocks don't do anything. In one sense they don't hurt, but a) it makes
auditing `unsafe` harder b) I'm not sure all "would be `unsafe` if not
in an `unsafe` parts" are marked `unsafe`, so it might be slightly
misleading.