Register allocators can't store an unsparse allocation mapping.
The design assumes that *every* instruction defines a local variable,
and that therefore allocations can be compacted without gaps into a
indexed vector backing store.

The premise simply isn't true.

Potentional fixes:

 - Change the backing vector to store `Option<LocalAlloc>` and put a
   `None` gap for instructions that don't define anything.

 - Use a data structure that allows us to store entries sparsely.

The former is wasteful (lots of `None`s) and error prone (every
instruction must inform the register allocator, even if it defines
nothing), so I've opted for the latter. A hashmap.