Add a variant of Module::push() that returns an operand.
It's called Module::push_and_make_operand().

This is useful when you are pushing an instruction into the module, and
you want to use the value it defines as an operand for a subsequent
instruction.

This is an idiom used a lot in testing. Currently we "magic up" operands
by "just knowing" the instruction index to put inside the operand. This
change removes the magic/guessing aspect.

This is a second attempt at a "quality of life" change I proposed in an
earlier PR, but was rejected because it encouraged swallowing errors:

https://github.com/ykjit/yk/pull/1006#discussion_r1511143181

This design does not encourage swallowing errors: if you don't want an
Operand back when you are adding an instruction to the module, then you
use push() instead of push_and_make_operand().