Don't run all tests multiple times on all tracers.
There is a tension here: we'd love to test everything multiple times but
then CI takes too long. CI has recently gone up to almost exactly an
hour, which is long enough to feel painful.

This commit tries to reduce some aspects of testing that somewhat
duplicate other stuff. This is a judgment call and there's never going
to be a perfect answer.