Use cached version of ykllvm in CI if it exists.
Our CI is getting too slow: about 57 minutes. Part of this is because we
build multiple copies of ykllvm, each of which takes multiple minutes.

This commit uses a cached version of ykllvm mounted in
`/opt/ykllvm_cache`, if it exists, to speed up CI. If a suitable cached
version is not found, we keep building our own copies in `.buildbot.sh`.
In other words, you can still use `.buildbot.sh` without knowing
anything about caching: it's just a useful optional extra.