Deal gracefully with trace buffer overflow.
There are two ways that we can detect overflow with PT, and previously
we only dealt gracefully with one of them. That very occasionally caused
the `trace_too_long.c` test to panic due to the `Some(Err(_)) =>
todo!()` you can see in this diff. This commit deals with that case,
which hopefully will mean that test now always succeeds.