Stringifying the IR shouldn't return a `CompilationError`.
It's not clear to me that any of the stringification can fail: even if
it does, it's not meaningfully a `CompilationError` in the sense that
enum is intended to be used. By using `CompilationError` we then muddy
up what's actually "compilation" (and which could error for possibly
reasonable reasons) from "other things".

This commit uses the more standard `Box<dyn Error>` idiom, though I
think it might be better to make all of these functions infallible: if
they fail, I think the system has broken an internal invariant and it
would be better to panic.