Remove the concept of temporary/unrecoverable compilation errors.
This simplifies the API because all the remaining `CompilationError`s
simply mean "we couldn't compile this trace". The wider system can then
retry tracing a location and might get a slightly different trace that
is compilable. Thus, in a sense, a `CompilationError` really signifies
"this is worth retrying".

In the future I can imagine that we might want to attach extra debugging
information to `CompilationError` (e.g. what IR element(s) contributed
to it?), but for now a simple `String` seems appropriate.