When `yk_testing` is enabled, make grabbing a Location lock deterministic.
In normal use, when we try and grab a `Location`s lock, we give up
entirely if it looks like it will take too long to grab the lock. This
solves a thundering herd problem that at least one of mt's unit tests
highlights.

However, it means that it is then non-deterministic when we encounter a
`Compiled` `Location` as to whether we fall back to the interpreter or
not. All of our C tests could technically have fallen pray to this, but
because we set `YKD_SERIALISED=1` and only used a single thread we
(mostly?) got away with it on x86. However the new event based framework
makes it much more likely that we notice this problem, because now we
genuinely are using multiple threads

This commit uses the `yk_testing` feature as a proxy for "make grabbing
a `Location`s lock deterministic".