Only use "always lock" when `yk_testing` is set.
`mt.rs`'s unit tests need to test the "real-world sometimes-can-fail"
locking, but ff72e72f accidentally caused those unit tests with the same
"always lock" code as when `yk_testing` is set. This commit fixes that.

Happily this commit also restores the speed of mt's unit tests. In
particular, the `only_one_thread_starts_tracing` test is hugely slow
when `yk_testing` is enabled, taking up to 80-90s to run on my machine,
instead of the 2-3 when `yk_testing` is disabled.