Don't write `YKD_STATS` output to a file called 1.
This tests needs to turn `YKD_STATS` on for the event based logging to
work, and it's better if that output doesn't appear on stderr, as it
muddies the test. However `YKD_STATS=1` writes to a file called `1`!
This commit changes it so that `YKD_STATS` output is written to
`/dev/null`, satisfying both constraints.