Make the state machine encoded in `MTThread` explicit.
Previously `MTThread` was effectively:

```rust
pub(crate) struct MTThread {
  running_trace: RefCell<Option<Arc<dyn CompiledTrace>>>,
  tracing: RefCell<Option<Arc<Mutex<HotLocation>>>>,
  thread_tracer: RefCell<Option<Box<dyn TraceRecorder>>>,
  promotions: RefCell<Option<Vec<usize>>>,
```

It's far from obvious but these 4 fields encode a state machine about
the current thread's state: and there are 3 separate states in that
state machine! This was far from obvious to me when I started: I spent
quite a while confused about the exact relationship between these fields
before realising this.

This commit breaks the statemachine out into an `enum` which is then
stored in a field in `MTThread`:

```rust
enum MTThreadState {
    Interpreting,
    Tracing {
        hl: Arc<Mutex<HotLocation>>,
        thread_tracer: Box<dyn TraceRecorder>,
        promotions: Vec<usize>,
    },
    Executing(Arc<dyn CompiledTrace>),
}
```

This was quite a bit more painful to do than the diff might suggest.

On the plus side, for example, we'd accidentally duplicated the
`HotLocation` in one part of the code, so we had two different variables
storing references to the same `HotLocation`. This commit fixes that.

On the negative side, we've forced a bit more responsibility onto
whoever calls a `transition_*` function. I think I might spy a future
opportunity to simplify some of this stuff, but that's something to
tackle in the future. This commit already introduces plenty enough
churn!