Move `SERIALISE_COMPILATION` check.
Instead of having three parts of the code (one of which -- the `use` --
only works by accident!) related to `SERIALISE_COMPILATION`, this commit
simplifies things to one part of the code. This is, of course, now
somewhat slower than before as we have to do an env var lookup, but given that:

  1. the common case is `SERIALISE_COMPILATION=1`,
  2. we then spawn a thread, which is probably orders of magnitude
     slower than an env var lookup,

I'm not particularly worried about the performance difference.