Add associated method `MTThread::with`.
Previously we exposed across the whole crate the `THREAD_MTTHREAD`
thread local, and gradually more parts of the system have made use of
that. The problem with that is that it looks weird (I keep thinking "why
the capital letters?" when reading the code) and

This commit adds an associated method `MTThread::with` such that:

```rust
THREAD_MTTHREAD.with(|mtt| ...)
```

can be written as:

```rust
MTThread::with(|mtt| ...)
```

This is a minor change, but it feels more idiomatic: it also allows us
to hide the actual thread local (which is now private rather than
`pub(crate)`), which hopefully will give us more flexibility in the
future.