Differentiate classes of compilation errors.
Previously `CompilationError` was a catch-all. I briefly thought about
making it just `dyn Error` or even just `String` before realising that
there are different things that we want to do depending on the error
that occurs.

This commit acknowledges that bugs/problems can happen in three
different parts/times: in the interpreter; in yk; or in the user's
environment.

The key part of this commit is to change `CompilationError` to:

```rust
pub(crate) enum CompilationError {
    /// Compilation failed for reasons that might be of interest to a programmer augmenting an
    /// interpreter with yk but not to the end user running a program on the interpreter.
    General(String),
    /// Something went wrong when compiling that is probably the result of a bug in yk.
    InternalError(String),
    /// A limit was exceeded (e.g. a pointer add that went beyond a struct). We try and check for
    /// as many of these as possible at compile time, but some can only be detected at JIT time.
    /// Most, perhaps all, of these suggest bugs in the interpreter.
    LimitExceeded(String),
    /// Compilation failed because an external resource was exhausted: the end user running the
    /// interpreter probably wants to be informed of this.
    ResourceExhausted(Box<dyn Error>),
}
```

Depending on which of those occur we then do different things at
run-time:

  * `General` | `LimitExceeded`: only warn if `yk_jitstate_debug`.
  * `InternalError`: warn (if not `yk_jitstate_debug`) or panic (if
    `yk_jitstate_debug`).
  * `ResourceExhausted`: warn.