Move the `RegisterAllocator`, don't pass `&mut`.
Passing `&mut` in this way has a number of second order effects which
are best avoided if they're not necessary.

In particular, it was not really possible to construct a nice
`default_codegen` method, because the ownership of the register
allocator interfered with things. This commit both enables that and also
adds `default_codegen`!