Minimise `unsafe` code.
Looking at `__yk_deopt` it was quite hard to work out what code was
`unsafe` and what wasn't. There is an unfortunate, and not really
resolvable, tension here: in a sense the function really is `unsafe`,
but in so marking it, all its contents are implicitly `unsafe`, which
makes it hard for us to focus on the bits that are really `unsafe`.

On balance, I tend to think that a function called from asm/C that takes
pointers is "obviously" unsafe externally, and that it's better for us
to be clearer internally about what is/isn't unsafe.