Change `symbol_vaddr` to `symbol_to_ptr`.
`dlsym` returns `*void` so having `symbol_vaddr(...) -> usize` feels weird,
because `usize` has another distinct meaning (as the Rust equivalent of
`size_t`). It also has complication on platforms where "virtual address
and pointer" are not interchangeable (e.g. CHERI).

This commit changes the function to (more-or-less) `symbol_to_ptr(...)
-> *const ()` which is the Rust equivalent of `*void`. This also made it
clearer to me that (with only one exception) it's easier if we take in
`&str` rather than `&CStr`, which this commit does at the same time.

Overall there is a small functional change here, but on x86, the change
is rather small. Using the (not yet stabilised...) `addr()` function
also makes clearer what's going on when we really want just the address
(e.g. for pretty printing).