Rename the `ykd_jitstate_debug` feature to `ykd`.
We already have `YKD_PRINT_JITSTATE` and `YKD_STATS`: we're likely to
gain more such features. Although `PRINT_JITSTATE` is mostly intended
for internal use, there will definitely be times when interpreter
authors will want to turn it on to understand aspects of performance.

This commit thus makes a move to homogenise the compile-time inclusion
of these features: `ykd`. It's off by default currently, although it
might be worth including it by default.