Make the AOT IR "static".
Currently an AOT IR instruction is an opcode and a collection of generic
operands. Whilst this is easy to lower to, it isn't easy to work with
beyond that. You have to know which operand index is used for what.

This change makes an AOT IR instruction an enum with named, specialised
fields. This is more type-safe and should be easier to work with.