Cope with an obviously broken cached ykllvm.
This hasn't happened yet, but will one day: our cached ykllvm will fail
to run (e.g. because of shared object problems) and will cause a yk PR
to fail in surprising ways.

This commit tries to spot a broken cached ykllvm, and builds its own
version. We hope that a later part of the CI chain will then build a
working ykllvm and put it in the cache.