Build ykllvm with debug info when `debug_assertions` is enabled.
We sometimes experience ykllvm crashes, but at the moment the resulting
stack traces don't come with line numbers. This commit enables debug
info in ykllvm when yk is compiled with `debug_assertions`. Given that
we also add ykllvm's assertions in this mode, any penalty from adding
the debug info is probably minor.